Updates to support Scala 2.13 (#289)

+++ modules/common/src/main/scala/surge/kafka/streams/KafkaStreamLifeCycleManagement.scala
@@ -22,7 +22,7 @@ trait KafkaStreamLifeCycleManagement[K, V, T <: KafkaStreamsConsumer[K, V], SV] 
   var lastConsumerSeen: Option[T] = None
   val healthCheckName: String
 
-  log.debug(s"Kafka streams ${settings.storeName} cache memory being used is {} KiB", Math.round(settings.cacheMemory / 1024))
+  log.debug(s"Kafka streams ${settings.storeName} cache memory being used is {} KiB", Math.round(settings.cacheMemory.toFloat / 1024f))
 
   protected val streamsConfig: Map[String, String]
   protected val metrics: Metrics
